feat: add ref for ImageSchema.ts
