+++ server/package.json
@@ -21,6 +21,7 @@
     "@types/exif": "^0.6.3",
     "@types/express": "^4.17.13",
     "@types/imagemagick": "^0.0.31",
+    "@types/mongoose": "^5.11.97",
     "@types/node": "^16.9.1",
     "@types/passport": "^1.0.7",
     "@types/yaml": "^1.9.7",
