+++ server/src/saveInDb/saveInDb.ts
@@ -3,8 +3,8 @@ import {logger} from "../logger/logger.js";
 import {readdir} from "fs/promises";
 import {__pathToGallery} from "../gallery/pathToGallery.js";
 import {fileMetadataAsync} from 'file-metadata';
-import {imageModel} from "../DbModels/Models.js";
 import {getUserIdFromToken} from "../getUserIdFromToken/getUserIdFromToken.js";
+import {imageModel} from "../DbModels/ImageSchema.js";
 
 /*
  * work after user request on upload file to the server
